Implement parallel job execution

+++ job_runner.h
@@ -4,6 +4,6 @@
 struct graph;
 struct dict;
 
-int run_jobs(struct graph *graph, struct dict *macros, int max_jobs);
+int run_jobs(struct graph *graph, struct dict *macros, size_t max_jobs);
 
 #endif
